requirements.txt
-i https://pypi.org/simple
antlr4-python3-runtime==4.9.3
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
idna==3.7; python_version >= '3.5'
importlib-metadata==7.1.0; python_version >= '3.8'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
packaging==24.0; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-arango==7.9.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
referencing==0.35.1; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.18.1; python_version >= '3.8'
setuptools==69.5.1; python_version >= '3.8'
simplejson==3.19.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stix2==3.0.1; python_version >= '3.6'
stix2-patterns==2.0.0; python_version >= '3.6'
tqdm==4.66.4; python_version >= '3.7'
urllib3==2.2.1; python_version >= '3.8'
zipp==3.18.2; python_version >= '3.8'
https://github.com/muchdogesec/stix2arango/archive/cti_processor_merge.zip